+++ src/app/latte-chat/mypage/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import MypageContainer from './MypageContainer'
+import MypageContainer from '@/features/user-mypage/containers/MypageContainer'
 
 export default function MypagePage() {
   return <MypageContainer />
